// src/worksfine_pb.ts
// @generated by protoc-gen-es v2.3.0 with parameter "target=ts,json_types=true"
// @generated from file worksfine.proto (package worksfine, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import { file_buf_validate_validate } from "./buf/validate/validate_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file worksfine.proto.
 */
export const file_worksfine: GenFile = /*@__PURE__*/
  fileDesc("Cg93b3Jrc2ZpbmUucHJvdG8SCXdvcmtzZmluZSLAAQoGUGVyc29uEhQKBG5hbWUYASABKAlCBrpIA8gBARIyCgdhZGRyZXNzGAIgASgLMhkud29ya3NmaW5lLlBlcnNvbi5BZGRyZXNzQga6SAPIAQEabAoHQWRkcmVzcxIXCgdhZGRyZXNzGAEgASgJQga6SAPIAQESFAoEY2l0eRgCIAEoCUIGukgDyAEBEhUKBXN0YXRlGAMgASgJQga6SAPIAQESGwoLcG9zdGFsX2NvZGUYBCABKAlCBrpIA8gBAWIGcHJvdG8z", [file_buf_validate_validate]);

/**
 * @generated from message worksfine.Person
 */
export type Person = Message<"worksfine.Person"> & {
  /**
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * @generated from field: worksfine.Person.Address address = 2;
   */
  address?: Person_Address;
};

/**
 * @generated from message worksfine.Person
 */
export type PersonJson = {
  /**
   * @generated from field: string name = 1;
   */
  name?: string;

  /**
   * @generated from field: worksfine.Person.Address address = 2;
   */
  address?: Person_AddressJson;
};

/**
 * Describes the message worksfine.Person.
 * Use `create(PersonSchema)` to create a new message.
 */
export const PersonSchema: GenMessage<Person, PersonJson> = /*@__PURE__*/
  messageDesc(file_worksfine, 0);

/**
 * @generated from message worksfine.Person.Address
 */
export type Person_Address = Message<"worksfine.Person.Address"> & {
  /**
   * @generated from field: string address = 1;
   */
  address: string;

  /**
   * @generated from field: string city = 2;
   */
  city: string;

  /**
   * @generated from field: string state = 3;
   */
  state: string;

  /**
   * @generated from field: string postal_code = 4;
   */
  postalCode: string;
};

/**
 * @generated from message worksfine.Person.Address
 */
export type Person_AddressJson = {
  /**
   * @generated from field: string address = 1;
   */
  address?: string;

  /**
   * @generated from field: string city = 2;
   */
  city?: string;

  /**
   * @generated from field: string state = 3;
   */
  state?: string;

  /**
   * @generated from field: string postal_code = 4;
   */
  postalCode?: string;
};

/**
 * Describes the message worksfine.Person.Address.
 * Use `create(Person_AddressSchema)` to create a new message.
 */
export const Person_AddressSchema: GenMessage<Person_Address, Person_AddressJson> = /*@__PURE__*/
  messageDesc(file_worksfine, 0, 0);